+++ src/components/Community/Search/History/HistoryList.scss
