+++ src/components/Community/Search/Search.scss
@@ -3,7 +3,7 @@
 .community-search-container {
   display: flex;
   height: 70px;
-  padding: 0 25px; 
+  padding: 0 25px;
   border-radius: 10px 10px 0 0;
 }
 
@@ -16,9 +16,64 @@
   height: 100%;
   background: rgba(0, 0, 0, 0.4);
 }
-  
+
 .community-result-background {
   width: 100%;
   height: 380px;
-  background-color: #FFFFFF;
+  background-color: #ffffff;
+}
+
+.search-input {
+  display: flex;
+  position: relative;
+  z-index: 5;
+  align-items: center;
+  flex-grow: 1;
+
+  & label {
+    position: relative;
+    z-index: 10;
+    top: 30px;
+
+    min-height: 17px;
+    align-self: flex-start;
+    margin-left: 30px;
+
+    color: #878888;
+  }
+
+  & input {
+    width: 100%;
+    max-width: 930px;
+    height: 43px;
+    border-radius: 100px;
+    margin-right: -30px;
+    outline: none;
+
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 17px;
+  }
+}
+
+.community-search-history {
+  width: 100%;
+  padding: 30px 10px 0;
+
+  font-weight: 500;
+  font-size: 14px;
+  line-height: 23px;
+  color: #252525;
+}
+
+.community-search-label {
+  span {
+    margin: 0 0 0 5px;
+  }
+}
+
+.search-icon {
+  color: $background-yellow;
+  margin-right: 10px;
+  margin-bottom: -1.3px;
 }
