+++ src/components/Community/hooks/useArticles.js
@@ -2,8 +2,10 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from 'lib/reactQuery/queryKeys';
 import { getArticleList } from 'lib/APIs/article';
 import { useSearchParams } from 'react-router-dom';
+import { useState } from 'react';
 
 export const useArticles = () => {
+  const [searchValue, setSearchValue] = useState('');
   const [searchParams, setSearchParams] = useSearchParams();
   const { Article } = QUERY_KEY;
 
@@ -13,13 +15,26 @@ export const useArticles = () => {
   const priority = searchParams.get('priority');
 
   const { fetchNextPage, hasNextPage, data } = useInfiniteQuery({
-    queryKey: [Article, { tag, firstPlace, secondPlace, priority }],
+    queryKey: [
+      Article,
+      { tag, firstPlace, secondPlace, priority, searchValue },
+    ],
     queryFn: ({ lastArticleId = 0 }) =>
-      getArticleList(tag, priority, firstPlace, secondPlace, lastArticleId, 10),
-    getNextPageParam: lastPage =>
-      lastPage?.data?.data.hasNext
-        ? lastPage.data.data.lastArticleId + 9
-        : undefined,
+      getArticleList(
+        tag,
+        priority,
+        firstPlace,
+        secondPlace,
+        lastArticleId,
+        searchValue,
+        10,
+      ),
+    getNextPageParam: lastPage => {
+      // console.log(lastPage.data.data.lastArticleId);
+      return lastPage?.data?.data.hasNext
+        ? Number(lastPage.data.data.lastArticleId) + 9
+        : undefined;
+    },
     enabled: !!priority,
   });
 
@@ -29,5 +44,6 @@ export const useArticles = () => {
     fetchNextPage,
     hasNextPage,
     pages,
+    setSearchValue,
   };
 };
