+++ src/components/common/hooks/useInput.js
@@ -2,15 +2,11 @@ import { useState } from 'react';
 
 export function useInput(validation = () => true) {
   const [value, setValue] = useState('');
-  const [focus, setFocus] = useState(false);
   const [valid, setValid] = useState(false);
 
   const handleChange = e => {
     setValue(e.target.value);
   };
-
-  const isFocus = () => setFocus(true);
-  const isBlur = () => setFocus(false);
 
   const isValid = () => setValid(validation(value));
 
@@ -18,9 +14,6 @@ export function useInput(validation = () => true) {
     value,
     setValue,
     handleChange,
-    focus,
-    isFocus,
-    isBlur,
     valid,
     isValid,
   };
